python_requirements.txt
flask
flask-sqlalchemy
flask-login
flask_bcrypt
waitress
configparser
logging
pyyaml
coloredlogs
flask_wtf